[YWA-809] Adequate presence policy behaviour
  Remove of the creation of the 7th day of the policy so there will be no more default 7 week long policies

  Allow delete on cascade of Time entries and presence days when a Presence policy is delete or delete on cascade of time entries
  if a presence day is deleted.

  Only allow deletion or update  of a presence policy or edition , creation ,delete of any dependent object of a presence policy ( presence days and time entries)
  if the presence policy has not been assigned to any employee.

+++ app/controllers/api/v1/time_entries_controller.rb
@@ -16,35 +16,48 @@ module API
         verified_dry_params(dry_validation_schema) do |attributes|
           resource = resources.new(time_entry_params(attributes))
           authorize! :create, resource
-
-          transactions do
-            resource.save!
-            update_affected_balances(resource.presence_day.presence_policy)
+          if is_resource_unlocked(presence_day)
+            transactions do
+              resource.save!
+            end
+            render_resource(resource, status: :created)
+          else
+            locked_error
           end
-          render_resource(resource, status: :created)
         end
       end
 
       def update
         verified_dry_params(dry_validation_schema) do |attributes|
-          transactions do
-            resource.update!(attributes)
-            update_affected_balances(resource.presence_day.presence_policy)
+          if is_resource_unlocked(resource.presence_day)
+            transactions do
+              resource.update!(attributes)
+            end
+            render_no_content
+          else
+            locked_error
           end
-          render_no_content
         end
       end
 
       def destroy
-        transactions do
-          update_affected_balances(resource.presence_day.presence_policy)
-          resource.destroy!
-          resource.presence_day.update_minutes!
+        if is_resource_unlocked(resource.presence_day)
+          transactions do
+            resource.destroy!
+            resource.presence_day.update_minutes!
+          end
+          render_no_content
+        else
+          locked_error
         end
-        render_no_content
       end
 
       private
+
+      def is_resource_unlocked(presence_day)
+        return true if presence_day.presence_policy.employees.empty?
+        false
+      end
 
       def resource
         @resource ||= Account.current.time_entries.find(params[:id])
